move verify clusters to its own package
